Tests for the meta net task
