+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor.rs
@@ -13,7 +13,6 @@ use crate::*;
 use futures::future::FutureExt;
 use futures::stream::StreamExt;
 use kitsune_p2p_fetch::*;
-use kitsune_p2p_timestamp::Timestamp;
 use kitsune_p2p_types::agent_info::AgentInfoSigned;
 use kitsune_p2p_types::async_lazy::AsyncLazy;
 use kitsune_p2p_types::tx2::tx2_api::*;
@@ -29,6 +28,7 @@ mod discover;
 pub(crate) mod meta_net;
 use meta_net::*;
 mod fetch;
+mod meta_net_task;
 mod space;
 use ghost_actor::dependencies::tracing;
 use space::*;
@@ -176,464 +176,21 @@ impl KitsuneP2pActor {
 
         let i_s = internal_sender.clone();
 
-        tokio::task::spawn({
-            let evt_sender = evt_sender.clone();
-            let host = host.clone();
-            let tuning_params = config.tuning_params.clone();
-            let fetch_pool = fetch_pool.clone();
-            async move {
-                let fetch_response_queue = &fetch_response_queue;
-                let fetch_pool = &fetch_pool;
-                ep_evt
-                    .for_each_concurrent(tuning_params.concurrent_limit_per_thread, move |event| {
-                        let evt_sender = evt_sender.clone();
-                        let host = host.clone();
-                        let i_s = i_s.clone();
-
-                        async move {
-                            let evt_sender = &evt_sender;
-
-                            match event {
-                                MetaNetEvt::Connected { remote_url, con } => {
-                                    let _ = i_s.new_con(remote_url, con.clone()).await;
-                                }
-                                MetaNetEvt::Disconnected { remote_url, con: _ } => {
-                                    let _ = i_s.del_con(remote_url).await;
-                                }
-                                MetaNetEvt::Request {
-                                    remote_url: _,
-                                    con,
-                                    data,
-                                    respond,
-                                } => {
-                                    match nodespace_is_authorized(
-                                        &host,
-                                        con.peer_id(),
-                                        data.maybe_space(),
-                                        Timestamp::now(),
-                                    )
-                                    .await
-                                    {
-                                        MetaNetAuth::UnauthorizedIgnore => {}
-                                        MetaNetAuth::UnauthorizedDisconnect => {
-                                            con.close(
-                                                UNAUTHORIZED_DISCONNECT_CODE,
-                                                UNAUTHORIZED_DISCONNECT_REASON,
-                                            )
-                                            .await;
-                                        }
-                                        MetaNetAuth::Authorized => {
-                                            match data {
-                                                wire::Wire::Call(wire::Call {
-                                                    space,
-                                                    to_agent,
-                                                    data,
-                                                    ..
-                                                }) => {
-                                                    let res = match evt_sender
-                                                        .call(space, to_agent, data.into())
-                                                        .await
-                                                    {
-                                                        Err(err) => {
-                                                            let reason = format!("{:?}", err);
-                                                            let fail = wire::Wire::failure(reason);
-                                                            respond(fail).await;
-                                                            return;
-                                                        }
-                                                        Ok(r) => r,
-                                                    };
-                                                    let resp = wire::Wire::call_resp(res.into());
-                                                    respond(resp).await;
-                                                }
-                                                wire::Wire::PeerGet(wire::PeerGet {
-                                                    space,
-                                                    agent,
-                                                }) => {
-                                                    let resp = match host
-                                                        .get_agent_info_signed(
-                                                            GetAgentInfoSignedEvt { space, agent },
-                                                        )
-                                                        .await
-                                                    {
-                                                        Ok(info) => wire::Wire::peer_get_resp(info),
-                                                        Err(err) => wire::Wire::failure(format!(
-                                                            "Error getting agent: {:?}",
-                                                            err,
-                                                        )),
-                                                    };
-                                                    respond(resp).await;
-                                                }
-                                                wire::Wire::PeerQuery(wire::PeerQuery {
-                                                    space,
-                                                    basis_loc,
-                                                }) => {
-                                                    // this *does* go over the network...
-                                                    // so we don't want it to be too many
-                                                    const LIMIT: u32 = 8;
-                                                    let query = QueryAgentsEvt::new(space)
-                                                        .near_basis(basis_loc)
-                                                        .limit(LIMIT);
-                                                    let resp = match evt_sender
-                                                        .query_agents(query)
-                                                        .await
-                                                    {
-                                                        Ok(list) => {
-                                                            wire::Wire::peer_query_resp(list)
-                                                        }
-                                                        Err(err) => wire::Wire::failure(format!(
-                                                            "Error querying agents: {:?}",
-                                                            err,
-                                                        )),
-                                                    };
-                                                    respond(resp).await;
-                                                }
-                                                data => unimplemented!("{:?}", data),
-                                            }
-                                        }
-                                    }
-                                }
-                                MetaNetEvt::Notify {
-                                    remote_url: url,
-                                    con,
-                                    data,
-                                } => {
-                                    match nodespace_is_authorized(
-                                        &host,
-                                        con.peer_id(),
-                                        data.maybe_space(),
-                                        Timestamp::now(),
-                                    )
-                                    .await
-                                    {
-                                        MetaNetAuth::UnauthorizedIgnore => {}
-                                        MetaNetAuth::UnauthorizedDisconnect => {
-                                            con.close(
-                                                UNAUTHORIZED_DISCONNECT_CODE,
-                                                UNAUTHORIZED_DISCONNECT_REASON,
-                                            )
-                                            .await;
-                                        }
-                                        MetaNetAuth::Authorized => {
-                                            match data {
-                                                wire::Wire::DelegateBroadcast(
-                                                    wire::DelegateBroadcast {
-                                                        space,
-                                                        basis,
-                                                        to_agent,
-                                                        mod_idx,
-                                                        mod_cnt,
-                                                        data,
-                                                    },
-                                                ) => match data {
-                                                    BroadcastData::Publish {
-                                                        source,
-                                                        op_hash_list,
-                                                        context,
-                                                    } => {
-                                                        if let Err(err) = i_s
-                                                            .incoming_publish(
-                                                                space,
-                                                                to_agent,
-                                                                source,
-                                                                op_hash_list,
-                                                                context,
-                                                                Some((basis, mod_idx, mod_cnt)),
-                                                            )
-                                                            .await
-                                                        {
-                                                            tracing::warn!(
-                                                    ?err,
-                                                    "failed to handle incoming delegate broadcast"
-                                                );
-                                                        }
-                                                    }
-                                                    data => {
-                                                        // one might be tempted to notify here
-                                                        // as in Broadcast below... but we
-                                                        // notify all relevant agents inside
-                                                        // the space incoming_delegate_broadcast
-                                                        // handler.
-                                                        if let Err(err) = i_s
-                                                            .incoming_delegate_broadcast(
-                                                                space, basis, to_agent, mod_idx,
-                                                                mod_cnt, data,
-                                                            )
-                                                            .await
-                                                        {
-                                                            tracing::warn!(
-                                                    ?err,
-                                                    "failed to handle incoming delegate broadcast"
-                                                );
-                                                        }
-                                                    }
-                                                },
-                                                wire::Wire::Broadcast(wire::Broadcast {
-                                                    space,
-                                                    to_agent,
-                                                    data,
-                                                    ..
-                                                }) => match data {
-                                                    BroadcastData::User(data) => {
-                                                        // TODO: Should we check if the basis is
-                                                        // held before calling notify?
-                                                        if let Err(err) = evt_sender
-                                                            .notify(space, to_agent, data)
-                                                            .await
-                                                        {
-                                                            tracing::warn!(
-                                                        ?err,
-                                                        "error processing incoming broadcast"
-                                                    );
-                                                        }
-                                                    }
-                                                    BroadcastData::AgentInfo(agent_info) => {
-                                                        // TODO: Should we check if the basis is
-                                                        // held before calling put_agent_info_signed?
-                                                        if let Err(err) = evt_sender
-                                                            .put_agent_info_signed(
-                                                                PutAgentInfoSignedEvt {
-                                                                    space,
-                                                                    peer_data: vec![agent_info],
-                                                                },
-                                                            )
-                                                            .await
-                                                        {
-                                                            tracing::warn!(
-                                                    ?err,
-                                                    "error processing incoming agent info broadcast"
-                                                );
-                                                        }
-                                                    }
-                                                    BroadcastData::Publish {
-                                                        source,
-                                                        op_hash_list,
-                                                        context,
-                                                    } => {
-                                                        if let Err(err) = i_s
-                                                            .incoming_publish(
-                                                                space,
-                                                                to_agent,
-                                                                source,
-                                                                op_hash_list,
-                                                                context,
-                                                                None,
-                                                            )
-                                                            .await
-                                                        {
-                                                            tracing::warn!(
-                                                        ?err,
-                                                        "failed to handle incoming broadcast"
-                                                    );
-                                                        }
-                                                    }
-                                                },
-                                                wire::Wire::Gossip(wire::Gossip {
-                                                    space,
-                                                    data,
-                                                    module,
-                                                }) => {
-                                                    let data: Vec<u8> = data.into();
-                                                    let data: Box<[u8]> = data.into_boxed_slice();
-                                                    if let Err(e) = i_s
-                                                        .incoming_gossip(
-                                                            space, con, url, data, module,
-                                                        )
-                                                        .await
-                                                    {
-                                                        tracing::warn!(
-                                                    "failed to handle incoming gossip: {:?}",
-                                                    e
-                                                );
-                                                    }
-                                                }
-                                                wire::Wire::FetchOp(wire::FetchOp {
-                                                    fetch_list,
-                                                }) => {
-                                                    for (space, key_list) in fetch_list {
-                                                        let mut hashes = Vec::new();
-                                                        let topo = match host
-                                                            .get_topology(space.clone())
-                                                            .await
-                                                        {
-                                                            Err(_) => continue,
-                                                            Ok(topo) => topo,
-                                                        };
-                                                        let mut regions = Vec::new();
-
-                                                        for key in key_list {
-                                                            match key {
-                                                                FetchKey::Region(region_coords) => {
-                                                                    regions.push((
-                                                                        region_coords,
-                                                                        region_coords
-                                                                            .to_bounds(&topo),
-                                                                    ));
-                                                                }
-                                                                FetchKey::Op(op_hash) => {
-                                                                    hashes.push(op_hash);
-                                                                }
-                                                            }
-                                                        }
-
-                                                        if !hashes.is_empty() {
-                                                            if let Ok(list) = evt_sender
-                                                        .fetch_op_data(FetchOpDataEvt {
-                                                            space: space.clone(),
-                                                            query: FetchOpDataEvtQuery::Hashes {
-                                                                op_hash_list: hashes,
-                                                                include_limbo: true,
-                                                            },
-                                                        })
-                                                        .await
-                                                    {
-                                                        for (_hash, op) in list {
-                                                            fetch_response_queue.enqueue_op(
-                                                                space.clone(),
-                                                                (con.clone(), url.clone(), None),
-                                                                op,
-                                                            );
-                                                        }
-                                                    }
-                                                        }
-
-                                                        for (coord, bound) in regions {
-                                                            if let Ok(list) = evt_sender
-                                                        .fetch_op_data(FetchOpDataEvt {
-                                                            space: space.clone(),
-                                                            query: FetchOpDataEvtQuery::Regions(
-                                                                vec![bound],
-                                                            ),
-                                                        })
-                                                        .await
-                                                    {
-                                                        let last_idx = list.len() - 1;
-                                                        for (idx, (_hash, op)) in
-                                                            list.into_iter().enumerate()
-                                                        {
-                                                            fetch_response_queue.enqueue_op(
-                                                                space.clone(),
-                                                                (
-                                                                    con.clone(),
-                                                                    url.clone(),
-                                                                    Some((coord, idx == last_idx)),
-                                                                ),
-                                                                op,
-                                                            );
-                                                        }
-                                                    }
-                                                        }
-                                                    }
-                                                }
-                                                wire::Wire::PushOpData(wire::PushOpData {
-                                                    op_data_list,
-                                                }) => {
-                                                    for (space, op_list) in op_data_list {
-                                                        for op in op_list {
-                                                            // hash the op
-                                                            let op_hash = match host
-                                                                .op_hash(op.op_data.clone())
-                                                                .await
-                                                            {
-                                                                Ok(op_hash) => op_hash,
-                                                                Err(_) => continue,
-                                                            };
-
-                                                            // trigger any delegation
-                                                            // that is pending on
-                                                            // having this data
-                                                            let _ = i_s
-                                                                .resolve_publish_pending_delegates(
-                                                                    space.clone(),
-                                                                    op_hash.clone(),
-                                                                )
-                                                                .await;
-
-                                                            // MAYBE: do something with the
-                                                            //        is_last bool?
-                                                            //        Right now we don't
-                                                            //        really care, because
-                                                            //        if it's a region
-                                                            //        we know it's gossip
-                                                            //        so it's okay if
-                                                            //        the context is
-                                                            //        `None`.
-                                                            let key =
-                                                                if let Some((region, _is_last)) =
-                                                                    op.region
-                                                                {
-                                                                    FetchKey::Region(region)
-                                                                } else {
-                                                                    FetchKey::Op(op_hash.clone())
-                                                                };
-                                                            let fetch_context = fetch_pool
-                                                                .remove(&key)
-                                                                .and_then(|i| i.context);
-
-                                                            // forward the received op
-                                                            let _ = evt_sender
-                                                                .receive_ops(
-                                                                    space.clone(),
-                                                                    vec![op.op_data],
-                                                                    fetch_context,
-                                                                )
-                                                                .await;
-                                                        }
-                                                    }
-                                                }
-                                                wire::Wire::MetricExchange(
-                                                    wire::MetricExchange { space, msgs },
-                                                ) => {
-                                                    let _ = i_s
-                                                        .incoming_metric_exchange(space, msgs)
-                                                        .await;
-                                                }
-                                                wire::Wire::PeerUnsolicited(
-                                                    wire::PeerUnsolicited { peer_list },
-                                                ) => {
-                                                    for peer in peer_list {
-                                                        if let Err(err) = evt_sender
-                                                        .put_agent_info_signed(
-                                                            PutAgentInfoSignedEvt {
-                                                                space: peer.space.clone(),
-                                                                peer_data: vec![peer.clone()],
-                                                            },
-                                                        ).await {
-                                                            tracing::warn!(?err, "error processing incoming agent info unsolicited");
-                                                        }
-                                                    }
-                                                }
-                                                wire::Wire::Failure(_)
-                                                | wire::Wire::Call(_)
-                                                | wire::Wire::CallResp(_)
-                                                | wire::Wire::PeerGet(_)
-                                                | wire::Wire::PeerGetResp(_)
-                                                | wire::Wire::PeerQuery(_)
-                                                | wire::Wire::PeerQueryResp(_) => {
-                                                    tracing::warn!(
-                                                        "received non-notify data in a notify"
-                                                    );
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    })
-                    .await;
-
-                tracing::error!(
-                    "KitsuneP2p: networking poll shutdown. Networking will no longer work!
-                You can ignore this is if it happened during node shutdown.
-                Otherwise please restart your node and report this error."
-                )
-            }
-        });
-
         let bandwidth_throttles = BandwidthThrottles::new(&config.tuning_params);
         let parallel_notify_permit = Arc::new(tokio::sync::Semaphore::new(
             config.tuning_params.concurrent_limit_per_thread,
         ));
+
+        MetaNetTask::new(
+            evt_sender.clone(),
+            host.clone(),
+            config.clone(),
+            fetch_pool.clone(),
+            fetch_response_queue,
+            ep_evt,
+            i_s,
+        )
+        .spawn();
 
         Ok(Self {
             channel_factory,
@@ -699,6 +256,7 @@ async fn create_meta_net(
 }
 
 use crate::spawn::actor::fetch::{FetchResponseConfig, FetchTask};
+use crate::spawn::actor::meta_net_task::MetaNetTask;
 use ghost_actor::dependencies::must_future::MustBoxFuture;
 
 impl ghost_actor::GhostControlHandler for KitsuneP2pActor {
