+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/test_util.rs
@@ -1,7 +1,9 @@
+mod host_event_stub;
 mod host_stub;
 mod internal_stub;
 mod space_internal_stub;
 
+pub use host_event_stub::HostEventStub;
 pub use host_stub::HostStub;
 pub use internal_stub::{InternalStub, InternalStubTest, InternalStubTestSender};
 pub use space_internal_stub::SpaceInternalStub;
